+++ NEWS.md
@@ -48,6 +48,7 @@ New library functions
 * `diskstat(path=pwd())` can be used to return statistics about the disk. ([#42248])
 * `copyuntil(out, io, delim)` and `copyline(out, io)` copy data into an `out::IO` stream ([#48273]).
 * `eachrsplit(string, pattern)` iterates split substrings right to left.
+* `split(string, dlm, ::Val{N})` splits to a `Tuple` generally without allocating.
 
 New library features
 --------------------